[task] 记录页主体功能

+++ MiHealth/app/src/main/java/com/hfut/mihealth/network/data/Food.kt
@@ -3,7 +3,7 @@ package com.hfut.mihealth.network.data
 data class Food(
     val foodid: Int,
     val name: String,
-    val calories: Double,
+    val calories: Int,
     val protein: Double,
     val carbs: Double,
     val fat: Double,
